+++ tests/scripts/refresh-decisions-lapi.php
@@ -42,7 +42,7 @@ $cacheRedisConfigs = [
     'redis_dsn' => 'redis://redis:6379',
 ];
 $redisCache = new Redis($cacheRedisConfigs, $logger);
-// Init CAPI remediation
+// Init LAPI remediation
 $remediationConfigs = [];
 $remediationEngine = new LapiRemediation($remediationConfigs, $lapiClient, $phpFileCache, $logger);
 // Retrieve fresh decisions from LAPI and update the cache